Moving form-related helpers from NavigationUITestContextExtensions to FormUITestContextExtensions
