+++ Lombiq.Tests.UI/Extensions/NavigationUITestContextExtensions.cs
@@ -5,7 +5,6 @@ using Lombiq.Tests.UI.Services;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Support.UI;
 using System;
-using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -159,36 +158,6 @@ namespace Lombiq.Tests.UI.Extensions
                 () => new WebDriverWait(context.Driver, TimeSpan.FromSeconds(10)).Until(
                     d => ((IJavaScriptExecutor)d).ExecuteScript("return document.readyState").Equals("complete")));
 
-        public static void SetDropdown<T>(this UITestContext context, string selectId, T value)
-            where T : Enum
-        {
-            context.ClickReliablyOn(By.Id(selectId));
-            context.Get(By.CssSelector(FormattableString.Invariant($"#{selectId} option[value='{(int)(object)value}']"))).Click();
-        }
-
-        public static void SetDropdownByText(this UITestContext context, string selectId, string value) =>
-            SetDropdownByText(context, By.Id(selectId), value);
-
-        public static void SetDropdownByText(this UITestContext context, By selectBy, string value)
-        {
-            context.ClickReliablyOn(selectBy);
-            context.Get(selectBy).Get(By.XPath($".//option[contains(., '{value}')]")).Click();
-        }
-
-        /// <summary>
-        /// Sets the value of the date picker via JavaScript and then raises the <c>change</c> event.
-        /// </summary>
-        public static void SetDatePicker(this UITestContext context, string id, DateTime value) =>
-            context.ExecuteScript(
-                FormattableString.Invariant($"document.getElementById('{id}').value = '{value:yyyy-MM-dd}';") +
-                $"document.getElementById('{id}').dispatchEvent(new Event('change'));");
-
-        public static DateTime GetDatePicker(this UITestContext context, string id) =>
-            DateTime.ParseExact(
-                context.Get(By.Id(id)).GetAttribute("value"),
-                "yyyy-MM-dd",
-                CultureInfo.InvariantCulture);
-
         /// <summary>
         /// A convenience method that merges <see cref="ElementRetrievalUITestContextExtensions.Get"/> and <see
         /// cref="NavigationWebElementExtensions.ClickReliably(IWebElement,UITestContext)"/> so the <paramref
@@ -207,74 +176,6 @@ namespace Lombiq.Tests.UI.Extensions
             TimeSpan? timeout = null,
             TimeSpan? interval = null) =>
             context.Get(by).ClickReliablyUntilPageLeave(context, timeout, interval);
-
-        /// <summary>
-        /// Finds the first submit button and clicks on it reliably.
-        /// </summary>
-        public static void ClickReliablyOnSubmit(this UITestContext context) =>
-            context.ClickReliablyOn(By.CssSelector("button[type='submit']"));
-
-        /// <summary>
-        /// Finds the "Add New" button.
-        /// </summary>
-        public static IWebElement GetAddNewButton(this UITestContext context) =>
-            context.Get(By.XPath("//button[contains(.,'Add New')]"));
-
-        /// <summary>
-        /// Opens the dropdown belonging to the "Add New" button. If <paramref name="byLocalMenuItem"/> is not <see
-        /// langword="null"/> it will click on that element within the dropdown context as well.
-        /// </summary>
-        public static void SelectAddNewDropdown(this UITestContext context, By byLocalMenuItem = null) =>
-            context.SelectFromBootstrapDropdownReliably(GetAddNewButton(context), byLocalMenuItem);
-
-        /// <summary>
-        /// Clicks on the <paramref name="dropdownButton"/> until the Bootstrap dropdown menu appears (up to 3 tries)
-        /// and then clicks on the <paramref name="byLocalMenuItem"/> within the dropdown menu's context.
-        /// </summary>
-        /// <param name="context">The current UI test context.</param>
-        /// <param name="dropdownButton">The button that reveals the Bootstrap dropdown menu.</param>
-        /// <param name="byLocalMenuItem">
-        /// The path inside the dropdown menu. If <see langword="null"/> then no selection (clicking) will be made, and
-        /// the dropdown is left open.
-        /// </param>
-        /// <exception cref="InvalidOperationException">
-        /// Thrown if clicking on the button didn't yield a dropdown menu even after retries.
-        /// </exception>
-        public static void SelectFromBootstrapDropdownReliably(
-            this UITestContext context,
-            IWebElement dropdownButton,
-            By byLocalMenuItem)
-        {
-            var byDropdownMenu = By.XPath("./following-sibling::*[contains(@class, 'dropdown-menu')]");
-
-            for (var i = 0; i < 3; i++)
-            {
-                dropdownButton.ClickReliably(context);
-
-                var dropdownMenu = dropdownButton.GetAll(byDropdownMenu).SingleOrDefault();
-                if (dropdownMenu != null)
-                {
-                    if (byLocalMenuItem != null) dropdownMenu.Get(byLocalMenuItem).ClickReliably(context);
-                    return;
-                }
-            }
-
-            throw new InvalidOperationException($"Couldn't open dropdown menu in 3 tries.");
-        }
-
-        /// <summary>
-        /// Clicks on the <paramref name="byDropdownButton"/> until the Bootstrap dropdown menu appears (up to 3 tries)
-        /// and then clicks on the menu item with the <paramref name="menuItemLinkText"/> within the dropdown menu's
-        /// context.
-        /// </summary>
-        /// <param name="context">The current UI test context.</param>
-        /// <param name="byDropdownButton">The path of the button that reveals the Bootstrap dropdown menu.</param>
-        /// <param name="menuItemLinkText">The text of the dropdown menu item.</param>
-        public static void SelectFromBootstrapDropdownReliably(
-            this UITestContext context,
-            By byDropdownButton,
-            string menuItemLinkText) =>
-            SelectFromBootstrapDropdownReliably(context, context.Get(byDropdownButton), By.LinkText(menuItemLinkText));
 
         /// <summary>
         /// Switches control to JS alert box, accepts it, and switches control back to main document or first frame.
